+++ client/store/cart.js
@@ -2,7 +2,7 @@ import axios from "axios";
 
 //action type
 const SET_CART = "SET_CART";
-const ADD_NEW_PRODUCTS_TO_CART = "ADD_NEW_PRODUCTS_TO_CART"
+//const ADD_NEW_PRODUCTS_TO_CART = "ADD_NEW_PRODUCTS_TO_CART"
 
 
 //action creator- fetch all products
@@ -14,12 +14,12 @@ const setCart = (cart) => {
 };
 
 // action creator - get cart with new products
-const addNewProductsToCart = (cart) => {
-  return {
-    type: ADD_NEW_PRODUCTS_TO_CART,
-    cart,
-  };
-};
+// const addNewProductsToCart = (cart) => {
+//   return {
+//     type: ADD_NEW_PRODUCTS_TO_CART,
+//     cart,
+//   };
+// };
 
 
 //fetch individual cart thunk
@@ -35,16 +35,16 @@ export const fetchCart = (userInfo) => {
   };
 };
 
-export const addingProductsToCart = (id) => {
-  return async (dispatch) => {
-    try {
-      const { data } = await axios.post(`/api/products/${id}`);
-      dispatch(addNewProductsToCart(data));
-    } catch (err) {
-      console.log(err);
-    }
-  };
-};
+// export const addingProductsToCart = (id) => {
+//   return async (dispatch) => {
+//     try {
+//       const { data } = await axios.post(`/api/products/${id}`);
+//       dispatch(addNewProductsToCart(data));
+//     } catch (err) {
+//       console.log(err);
+//     }
+//   };
+// };
 
 const initialState = [];
 
@@ -54,8 +54,8 @@ export default function cartReducer(state = initialState, action) {
   switch (action.type) {
     case SET_CART:
       return action.cart;
-    case ADD_NEW_PRODUCTS_TO_CART:
-      return [...state, action.cart ]
+    // case ADD_NEW_PRODUCTS_TO_CART:
+    //   return [...state, action.cart ]
     default:
       return state;
   }
